update folders into versions instead of /bad and /good
Using /bad and /good implies two version. Instead, sometimes
we may have multiple iterations of the same application which
will require us to see the various version.

This changes the set up to have version in the form `v0`, ..., `vn`,
and the goal is the final one that uses the design pattern will
be in `vfinal`.